+++ client/components/Animation.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Animation = () => {
+        return (
+     <div>   
+         <div class="outerCircle"></div>
+             <div class="innerCircle"></div>
+                <div class="icon"></div>     
+</div>
+        )
+    }
+    export default Animation     
+
