+++ client/components/App.jsx
@@ -6,7 +6,8 @@ import Header from "./Header";
 import ArticleList from "./ArticleList";
 import Footer from './Footer';
 import TranslateButtons from './TranslateButtons';
-
+import Animation from './Animation';
+import Overlay from './Overlay';
 // const googleNewsApi = 'https://newsapi.org/v2/top-headlines?country=nz&apiKey=4db317e841ff4a9ab8831f158ed48c29'
 
 class App extends React.Component {
@@ -34,6 +35,8 @@ class App extends React.Component {
     return (
       <React.Fragment>
         <Header />
+        <Overlay />
+        {/* <Animation /> */}
         <TranslateButtons/>
         {/* <Nav /> */}
         <ArticleList />
