+++ client/components/TranslateButtons.jsx
@@ -20,18 +20,20 @@ class TranslateButtons extends React.Component {
         return (
             <nav className="navbar navbar-expand-lg ">
                 <div className="collapse navbar-collapse" id="navbarNav">
+                 {/* <div id='Header'> */}
                     <ul className="navbar-nav">
                         <li className="nav-item">
-                            <button className="nav-link" onClick={() => this.handleLanguageChange(null)}>English</button>
+                            <button className="nav-link" onClick={() => this.handleLanguageChange(null)} type="button" class="btn btn-light btn-lg">English</button>
                         </li>
                         <li className="nav-item">
-                            <button className="nav-link" onClick={() => this.handleLanguageChange(gollumify)}>Gollumify</button>
+                            <button className="nav-link" onClick={() => this.handleLanguageChange(gollumify)} type="button" class="btn btn-light btn-lg">Gollumify</button>
                         </li>
                         <li className="nav-item">
-                            <button className="nav-link" onClick={() => this.handleLanguageChange(nazgulify)}>Nazgul</button>
+                            <button className="nav-link" onClick={() => this.handleLanguageChange(nazgulify)} type="button" class="btn btn-light btn-lg">Nazgul</button>
                         </li>
                     </ul>
                 </div>
+            {/* </div> */}
             </nav>
         )
     }
