+++ public/main.css
@@ -21,15 +21,16 @@
 
 }
 #list_articles {
-    padding-top: 6%;
+    /* padding-top: 6%; */
     letter-spacing: 1px;
     background-color: beige;
     overflow: auto;
     max-height: 700px;
+    max-width: 80%;
 }
 
 #app {
-    width: 60%;
+    width: 55%;
     margin: 0 auto;
     background-color: rgba(255,255,255,0.3);
     padding: 20px;
@@ -118,3 +119,115 @@ body {
 #myBtn:hover {
     background-color: #555; /* Add a dark-grey background on hover */
 }
+
+.icon {
+    background: url('images/hive.png') 0 0 no-repeat;
+    width: 76px;
+    height: 76px;
+    top: -200px;
+    margin: 0 auto;
+    position: relative;
+    }
+    .icon:hover {
+    -moz-animation: rotatecircle 4s infinite linear;
+    -webkit-animation: rotatecircle 4s linear;
+    }
+    .outerCircle {
+    background-color: transparent;
+    border: 8px solid rgba(97, 82, 72, 0.9);
+    opacity: .9;
+    border-right: 5px solid transparent;
+    border-left: 5px solid transparent;
+    border-radius: 100px;
+    width: 103px;
+    height: 103px;
+    margin: 0 auto;
+    -moz-animation: spinPulse 3s infinite ease-in-out;
+    -webkit-animation: spinPulse 3s infinite ease-in-out;
+    }
+    .innerCircle {
+    background-color: transparent;
+    border: 5px solid rgba(189, 215, 60, 0.6);
+    opacity: .9;
+    border-left: 5px solid transparent;
+    border-right: 5px solid transparent;
+    border-radius: 100px;
+    top: -110px;
+    width: 92px;
+    height: 92px;
+    margin: 0 auto;
+    position: relative;
+    -moz-animation: spinoffPulse 1s infinite linear;
+    -webkit-animation: spinoffPulse 1s infinite linear;
+    
+    }
+    
+    .animation {
+       
+    }
+    @-moz-keyframes spinPulse {
+        0% { -moz-transform:rotate(160deg); opacity:0; box-shadow:0 0 1px #bdd73c;}
+        50% { -moz-transform:rotate(145deg); opacity:1; }
+        100% { -moz-transform:rotate(-320deg); opacity:0; }
+    }
+    @-moz-keyframes spinoffPulse {
+        0% { -moz-transform:rotate(0deg); }
+        100% { -moz-transform:rotate(360deg);  }
+    }
+    @-webkit-keyframes spinPulse {
+        0% { -webkit-transform:rotate(160deg); opacity:0; box-shadow:0 0 1px #bdd73c; }
+        50% { -webkit-transform:rotate(145deg); opacity:1;}
+        100% { -webkit-transform:rotate(-320deg); opacity:0; }
+    }
+    @-webkit-keyframes spinoffPulse {
+        0% { -webkit-transform:rotate(0deg); }
+        100% { -webkit-transform:rotate(360deg); }
+    }
+    @-moz-keyframes rotatecircle {
+        0% { -moz-transform:rotate(0deg); }
+        100% { -moz-transform:rotate(-360deg); }
+    }
+    
+    @-webkit-keyframes rotatecircle {
+        0% { -webkit-transform:rotate(0deg); }
+    100% { -webkit-transform:rotate(-360deg); }
+    }
+
+    .src-image {
+        display: none;
+      }
+      
+      .card {
+        overflow: hidden;
+        position: relative;
+        border: 1px solid #CCC;
+        border-radius: 8px;
+        text-align: center;
+        padding: 0;
+        background-color: #284c79;
+        color: rgb(136, 172, 217);
+      }
+      
+      .card .header-bg {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 70px;
+        border-bottom: 1px blue;
+        z-index: 1;
+      }
+      .card .avatar {
+        position: relative;
+        margin-top: 15px;
+        z-index: 100;
+      }
+      
+      .card .avatar img {
+        width: 100px;
+        height: 100px;
+        -webkit-border-radius: 50%;
+        -moz-border-radius: 50%;
+        border-radius: 50%;
+        border: 5px solid rgba(0,0,30,0.8);
+      }
